Split dog20 tests
